requirements.txt
requests>=2.31.0
python-telegram-bot==21.9
aiohttp>=3.8.0
python-dotenv==1.0.0
requests>=2.31.0


